requirements.txt
Django==1.11

# authentication
django-allauth==0.32

# api
coreapi==2.3.3
djangorestframework==3.8.2
